refactor: add menu group repository
