+++ src/main/java/kitchenpos/domain/MenuGroup.java
@@ -1,8 +1,19 @@
 package kitchenpos.domain;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
 public class MenuGroup {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
+
+    protected MenuGroup() {
+    }
 
     public MenuGroup(String name) {
         this(null, name);
