+++ src/main/java/kitchenpos/domain/MenuGroupRepository.java
@@ -1,11 +1,11 @@
-package kitchenpos.dao;
+package kitchenpos.domain;
 
-import kitchenpos.domain.MenuGroup;
+import org.springframework.data.repository.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuGroupDao {
+public interface MenuGroupRepository extends Repository<MenuGroup, Long> {
     MenuGroup save(MenuGroup entity);
 
     Optional<MenuGroup> findById(Long id);
